pyproject.toml
[project]
name = "ipl-analytics"
version = "0.1.0"
description = ""
authors = [
    {name = "Hitesh Kumar",email = "coderhitesh1911@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=3.0.0,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.32.0,<1.0.0)",
    "pydantic-settings (>=2.6.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "ipl_analytics", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
